truss: Replace FreeBSD truss(1) with a homegrown version for NetBSD
The FreeBSD version of truss(1) has been adapted in a local tree
for building on NetBSD. However it had issues:

 - no tracing over a different ABI is planned at this point
 - no need for MD code on NetBSD (per-ABI)
 - handling of syscalls was specific for the FreeBSD kernel only
 - ptrace(2) interface to implement tracing was specific to FreeBSD
   kernel only

This means that almost all of the code was specific to FreeBSD only.

It was easier to start from scratch:

 - fork picotrace
 - extend common/ for new program needs
 - implement two modes of operation as distinct programs in a single
   executable binary: count (-c) and events tracing & reporting

This version of truss(1) for NetBSD implements all features of
FreeBSD truss(1) - as in, all the arguments passed to the program are
handled in the same way. Only a subest of syscalls has implemented
deserializing of arguments, this will be improved over time.
Right now no ioctl(2) operations are decoded on per-operation basis.

The output format from NetBSD truss(1) is similar to the one from
FreeBSD truss(1), however they differ in certain ways and keeping
exact compatibility here is out of scope as it is not practical.
The output is specific for FreeBSD syscalls, and NetBSD ships with
a different ones. However whenever it is fine, there is kept
compatibility with FreeBSD.

The NetBSD truss(1) version implements or plans to implement
few extra features:

 - auxv printing for exec events for a new command line option -x
 - SIGINFO handling of the counting mode (-c)

This version is work in progress.