picotrace: Switch from SPRINTF() + PRINTF() to StringList
Use stringlist(3) API for generating a report line. This removes
restriction on line length and allows better compiler diagnostics as
format and arguments passed to strings are now checked.

Negative aspect of this change is that sl API is more elaborated
and consists of more small allocations on heap that might have a
performance overhead.

No functional change intended in the program.